Add files via upload
Update of all the scripts. The correlators script has not changed since last time. The effective masses script now contains extra parts that calculates the effective masses of the operators analysis, when operators are removed or added to a basis. The operators analysis script was changed in the following way: Add function starts from a 2x2 matrix and starts adding operators to that basis performing the GEVP for each basis. It goes up to the original nr of operators minus 1. The remove function loops over each operator keep the nr constant but removing only that specific operator, so the size of the matrix is the original minus one all the time. The From list function takes a list you provide in the "files_ens.py" and it computes the new GEVP with that chosen set of operators. Every time you choose a new set, it will write a new dataset in the hdf5 file, it will not delete the previous data. 
The GEVP was also improved, now the process should converge better and show less problems. The eigenvalues are sorted by size, but one can change this by explicitly choosing other type of sorting. 
The plots are all consistent for the correlators and effective masses. All of them are saved automatically. The fit plots are still under construction as well as the fit script. The join plots probably will not work because I changed the names. But that will come. There are still some issues showing the best range of the data in the plots, and that I am still figuring it out.