Add files via upload
Updated version of the info of the ensemble X451. This one contains the list to add/remove arbitrary operators. One must write it with the name: "chosen_operators_list", which contains the operators that stay for the GEVP. If the list contains all the operators of the original set, then nothing is done.