Update of the scripts
Some bugs were fixed. There are two new variables that need to be defined in files_ens.py, it's called "the_hadron_state" and "the_non_interacting_levels", to prepare for the Ratio of Correlators improvement. The errors in the plots should appear okay now, and the variable "chosen_tmin" in the scripts "plot_fits_bnlah.py" and "plot_fitted_blah.py" corresponds to the tmin of choice for the fits. It must be the one one sees in the plot, no shift required. 
Maybe the Operators Analysis selection is slightly affected. The GEVP should work fine, because this was tested with the D200 and the X451. Maybe I made some mistakes with the spacing and/or forgot something, just let me know.